Move some code from libratom.lib.entities to libratom.lib.core
